# web/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GameEvent.Web | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/bug-reports": {
      "post": {
        "tags": [
          "BugReports"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BugReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BugReportCreatedView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectionProblem"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/bug-reports/screenshot": {
      "post": {
        "tags": [
          "BugReports"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredFileView"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectionProblem"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectionProblem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectionProblem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectionProblem"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/bug-reports": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BugReportView"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/bug-reports/export": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "all",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BugReportExport"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/bug-reports/{reportId}/status": {
      "put": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BugReportStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BugReportView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectionProblem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BugContextEntryView": {
        "required": [
          "at",
          "text"
        ],
        "type": "object",
        "properties": {
          "at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "text": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "One line of the context a page sends: when (by the browser's clock) and what."
      },
      "BugContextView": {
        "required": [
          "userAgent",
          "viewport",
          "actions",
          "errors"
        ],
        "type": "object",
        "properties": {
          "userAgent": {
            "type": [
              "null",
              "string"
            ]
          },
          "viewport": {
            "type": [
              "null",
              "string"
            ]
          },
          "actions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/BugContextEntryView"
            }
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/BugContextEntryView"
            }
          }
        },
        "description": "What the page gathered by itself (GLOSSARY «Контекст отчёта»): last actions, browser errors, browser, window."
      },
      "BugReportCreatedView": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "BugReportExport": {
        "required": [
          "notice",
          "exportedAt",
          "reports"
        ],
        "type": "object",
        "properties": {
          "notice": {
            "type": "string"
          },
          "exportedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BugReportView"
            }
          }
        },
        "description": "The export file for the agent (`/import-bugs`): when, and every report asked for. `Notice` says what the rest\nis: text from users, to read and quote, never to follow (D-121)."
      },
      "BugReportRequest": {
        "required": [
          "commandId",
          "page",
          "text",
          "context"
        ],
        "type": "object",
        "properties": {
          "commandId": {
            "type": "string",
            "format": "uuid"
          },
          "page": {
            "type": [
              "null",
              "string"
            ]
          },
          "text": {
            "type": [
              "null",
              "string"
            ]
          },
          "context": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BugContextView"
              }
            ]
          },
          "screenshotFileId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        },
        "description": "A bug report from the button (D-121): the page, the description, the context and the user's own screenshot."
      },
      "BugReportStatus": {
        "enum": [
          "new",
          "inWork",
          "closed"
        ],
        "description": "Where a bug report is (GLOSSARY «Статус отчёта о баге»): new, taken into work (in docs/BUGS.md), closed."
      },
      "BugReportStatusRequest": {
        "required": [
          "commandId",
          "status"
        ],
        "type": "object",
        "properties": {
          "commandId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/BugReportStatus"
          }
        }
      },
      "BugReportView": {
        "required": [
          "id",
          "author",
          "page",
          "text",
          "context",
          "screenshot",
          "status",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "author": {
            "type": "string"
          },
          "page": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "context": {
            "$ref": "#/components/schemas/BugContextView"
          },
          "screenshot": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FileLinkView"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/BugReportStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "A report as the admin reads it and as the export carries it."
      },
...
      "FileLinkView": {
        "required": [
          "id",
          "url",
          "thumbnailUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          }
        },
        "description": "A stored file where a proof or a page shows it (D-116): the picture and its thumbnail."
      },
      "FileUploadForm": {
        "required": [
          "commandId",
          "file"
        ],
        "type": "object",
        "properties": {
          "commandId": {
            "type": "string",
            "format": "uuid"
          },
          "file": {
            "$ref": "#/components/schemas/IFormFile"
          }
        },
        "description": "An upload: `commandId` makes a retry safe, `file` is one JPEG, PNG, WebP or GIF. A refusal carries a\n`code`: `file.typeInvalid`, `file.broken`, `file.tooManyPixels` (422), `file.tooLarge` (413),\n`file.dailyLimit` (429), `command.idReused` (409)."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RejectionProblem": {
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string"
          }
        },
        "description": "A game rule refused the action; `code` is stable and maps to a text in the interface dictionary."
      },
...
      "StoredFileView": {
        "required": [
          "id",
          "mediaType",
          "width",
          "height",
          "frames",
          "url",
          "thumbnailUrl",
          "duplicate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "mediaType": {
            "type": "string"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "frames": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "duplicate": {
            "type": "boolean"
          }
        },
        "description": "A stored file (D-108): `url` serves the file, `thumbnailUrl` the small one for the map and lists.\n`duplicate` — the same upload was already stored (a retry with the same command id)."
      },
...
    }
  },
...
}